# apps/admin/src/api/generated/cMSAdminAPI.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CMS Admin API
 * CMS 管理后台接口契约,前后端唯一事实源。
 *
 * 修改流程:改本文件 → `pnpm gen:api` 双端生成 → 双端实现。
 * 生成物(apps/admin/src/api/generated/、apps/server/gen/)禁止手改。
 *
 * 响应包装:传输层统一为 `{code, message, data}`(code 等于 HTTP 状态码);
 * 本契约描述的是 data 载荷,admin 在 mutator(src/api/client.ts)统一解包。
 *
 * OpenAPI spec version: 0.2.0
 */
export interface Error {
  /** HTTP 状态码 */
  code: number;
  /** 错误描述 */
  message: string;
}

export interface PageRequest {
  /** @minimum 1 */
  page?: number;
  /**
     * @minimum 1
     * @maximum 100
     */
  pageSize?: number;
}

export interface PageMeta {
  /** 总条数 */
  total: number;
}

export interface HealthzResponse {
  status: string;
}

export interface LoginRequest {
  /**
     * @minLength 1
     * @maxLength 64
     */
  username: string;
  /**
     * @minLength 6
     * @maxLength 64
     */
  password: string;
}

export interface UserInfo {
  id: number;
  username: string;
  nickname: string;
  email?: string;
  status: boolean;
  /** 角色码列表 */
  roles: string[];
  /** 权限码列表(阶段 2/3 填充) */
  permissions: string[];
}

export interface LoginResponse {
  token: string;
  /** token 过期时间 */
  expiresAt: string;
  user: UserInfo;
}

export interface ChangePasswordRequest {
  /**
     * @minLength 6
     * @maxLength 64
     */
  oldPassword: string;
  /**
     * @minLength 6
     * @maxLength 64
     */
  newPassword: string;
}

export interface UserItem {
  id: number;
  username: string;
  nickname: string;
  email?: string;
  status: boolean;
  /** 内置管理员,不可删除/禁用 */
  isBuiltin?: boolean;
  /** @nullable */
  lastLoginAt?: string | null;
  roleIds: number[];
}

export interface UserListResponse {
  list: UserItem[];
  total: number;
}

export interface UserCreateRequest {
  /**
     * @minLength 1
     * @maxLength 64
     */
  username: string;
  /**
     * @minLength 6
     * @maxLength 64
     */
  password: string;
  /** @maxLength 64 */
  nickname?: string;
  /** @maxLength 128 */
  email?: string;
  status?: boolean;
  roleIds?: number[];
}

export interface UserUpdateRequest {
  /** @maxLength 64 */
  nickname: string;
  /** @maxLength 128 */
  email?: string;
}

export interface StatusRequest {
  status: boolean;
}

export interface RoleIdsRequest {
  roleIds: number[];
}

export interface RoleItem {
  id: number;
  code: string;
  name: string;
  remark?: string;
  status: boolean;
  isBuiltin: boolean;
  /** 角色拥有的权限点 ID */
  permissionIds: number[];
}

export interface RoleListResponse {
  list: RoleItem[];
  total: number;
}

export interface RoleBrief {
  id: number;
  code: string;
  name: string;
  status: boolean;
}

export interface RoleRequest {
  /**
     * @minLength 1
     * @maxLength 64
     */
  code: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  /** @maxLength 255 */
  remark?: string;
  status?: boolean;
}

export interface PermissionIdsRequest {
  permissionIds: number[];
}

export type PermissionNodeType = typeof PermissionNodeType[keyof typeof PermissionNodeType];


export const PermissionNodeType = {
  menu: 'menu',
  api: 'api',
} as const;

export interface PermissionNode {
  id: number;
  code: string;
  name: string;
  type: PermissionNodeType;
  parentId: number;
  children: PermissionNode[];
}

export type OperationLogItemStatus = typeof OperationLogItemStatus[keyof typeof OperationLogItemStatus];


export const OperationLogItemStatus = {
  success: 'success',
  failed: 'failed',
} as const;

export interface OperationLogItem {
  id: number;
  userId: number;
  /** 操作人快照;登录失败记尝试的登录名 */
  username?: string;
  /** 资源.动作,如 user.delete */
  action: string;
  resource: string;
  resourceId?: string;
  /** 人话描述,如"删除用户 张三(zhangsan)" */
  description: string;
  status: OperationLogItemStatus;
  ip?: string;
  createdAt: string;
}

export interface OperationLogListResponse {
  list: OperationLogItem[];
  total: number;
}

export interface ConfigItem {
  key: string;
  /** 原样字符串,结构化数据自行 JSON 编码 */
  value: string;
  remark?: string;
}

export interface ConfigGroupResponse {
  group: string;
  items: ConfigItem[];
}

export interface ConfigUpdateRequest {
  items: ConfigItem[];
}

export interface Dict {
  id: number;
  code: string;
  name: string;
  remark?: string;
  status: boolean;
}

export interface DictUpsertRequest {
  /**
     * @minLength 1
     * @maxLength 64
     */
  code: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  /** @maxLength 255 */
  remark?: string;
  status?: boolean;
}

export interface DictEntry {
  id: number;
  dictId: number;
  label: string;
  value: string;
  sort: number;
  status: boolean;
}

export interface DictEntryUpsertRequest {
  /**
     * @minLength 1
     * @maxLength 64
     */
  label: string;
  /**
     * @minLength 1
     * @maxLength 64
     */
  value: string;
  sort?: number;
  status?: boolean;
}

export type PageParameter = number;

export type PageSizeParameter = number;

export type ListUsersParams = {
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParameter;
/**
 * 按用户名/昵称模糊匹配
 */
keyword?: string;
status?: boolean;
};

export type ListRolesParams = {
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParameter;
/**
 * 按角色名/编码模糊匹配
 */
keyword?: string;
status?: boolean;
};

export type ListOperationLogsParams = {
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: PageSizeParameter;
/**
 * 按操作人模糊匹配
 */
username?: string;
/**
 * 资源类型精确匹配,如 user / role
 */
resource?: string;
/**
 * 动作精确匹配,如 user.delete
 */
action?: string;
/**
 * 成功/失败筛选
 */
status?: ListOperationLogsStatus;
startTime?: string;
endTime?: string;
};

export type ListOperationLogsStatus = typeof ListOperationLogsStatus[keyof typeof ListOperationLogsStatus];


export const ListOperationLogsStatus = {
  success: 'success',
  failed: 'failed',
} as const;

export type ListDictsParams = {
keyword?: string;
};