# openapi.yaml
openapi: 3.0.3
info:
  title: CMS Admin API
  version: 0.2.0
  description: |
    CMS 管理后台接口契约,前后端唯一事实源。

    修改流程:改本文件 → `pnpm gen:api` 双端生成 → 双端实现。
    生成物(apps/admin/src/api/generated/、apps/server/gen/)禁止手改。

    响应包装:传输层统一为 `{code, message, data}`(code 等于 HTTP 状态码);
    本契约描述的是 data 载荷,admin 在 mutator(src/api/client.ts)统一解包。
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: 用户列表
      tags:
        - users
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema:
            type: string
          description: 按用户名/昵称模糊匹配
        - name: status
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: 用户分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      summary: 新建用户
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: 新建后的用户
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserItem"
        "409":
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}:
    get:
      operationId: getUser
      summary: 用户详情
      tags:
        - users
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 用户详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserItem"
    put:
      operationId: updateUser
      summary: 编辑用户基础信息
      tags:
        - users
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: 更新后的用户
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserItem"
    delete:
      operationId: deleteUser
      summary: 删除用户(不可删自己与内置管理员)
      tags:
        - users
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: 删除成功
        "403":
          description: 不可操作自己或内置管理员
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      summary: 启用/禁用用户(不可操作自己)
      tags:
        - users
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusRequest"
      responses:
        "204":
          description: 更新成功
        "403":
          description: 不可操作自己
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}/roles:
    put:
      operationId: updateUserRoles
      summary: 分配角色(全量覆盖)
      tags:
        - users
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleIdsRequest"
      responses:
        "204":
          description: 分配成功
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP 状态码
        message:
          type: string
          description: 错误描述
...
    UserItem:
      type: object
      required:
        - id
        - username
        - nickname
        - status
        - roleIds
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        nickname:
          type: string
        email:
          type: string
        status:
          type: boolean
        isBuiltin:
          type: boolean
          description: 内置管理员,不可删除/禁用
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        roleIds:
          type: array
          items:
            type: integer
            format: int64
    UserListResponse:
      type: object
      required:
        - list
        - total
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/UserItem"
        total:
          type: integer
    UserCreateRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          minLength: 6
          maxLength: 64
        nickname:
          type: string
          maxLength: 64
        email:
          type: string
          maxLength: 128
        status:
          type: boolean
          default: true
        roleIds:
          type: array
          items:
            type: integer
            format: int64
    UserUpdateRequest:
      type: object
      required:
        - nickname
      properties:
        nickname:
          type: string
          maxLength: 64
        email:
          type: string
          maxLength: 128
    StatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: boolean
    RoleIdsRequest:
      type: object
      required:
        - roleIds
      properties:
        roleIds:
          type: array
          items:
            type: integer
            format: int64
...